Update to using existing api-version with only additional properties so no breaking cchanges
