+++ server/src/stacks/2020-10-01/stacks/function-app-stacks/Node.ts
@@ -44,6 +44,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   netFrameworkVersion: 'v6.0',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node20EOL,
               },
               linuxRuntimeSettings: {
@@ -65,6 +72,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|20',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node20EOL,
               },
             },
@@ -99,6 +113,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   netFrameworkVersion: 'v6.0',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node18EOL,
               },
               linuxRuntimeSettings: {
@@ -120,6 +141,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|18',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node18EOL,
               },
             },
@@ -154,6 +182,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   netFrameworkVersion: 'v6.0',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node16EOL,
               },
               linuxRuntimeSettings: {
@@ -175,6 +210,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|16',
                 },
                 supportedFunctionsExtensionVersions: ['~4'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node16EOL,
               },
             },
@@ -208,6 +250,18 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   netFrameworkVersion: 'v6.0',
                 },
                 supportedFunctionsExtensionVersions: ['~4', '~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: false,
+                  },
+                ],
                 endOfLifeDate: node14EOL,
               },
               linuxRuntimeSettings: {
@@ -228,6 +282,18 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|14',
                 },
                 supportedFunctionsExtensionVersions: ['~4', '~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~4',
+                    isDeprecated: false,
+                    isDefault: true,
+                  },
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: false,
+                  },
+                ],
                 endOfLifeDate: node14EOL,
               },
             },
@@ -261,6 +327,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   use32BitWorkerProcess: true,
                 },
                 supportedFunctionsExtensionVersions: ['~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node12EOL,
               },
               linuxRuntimeSettings: {
@@ -282,6 +355,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|12',
                 },
                 supportedFunctionsExtensionVersions: ['~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node12EOL,
               },
             },
@@ -315,6 +395,18 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   use32BitWorkerProcess: true,
                 },
                 supportedFunctionsExtensionVersions: ['~2', '~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~2',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: false,
+                  },
+                ],
                 endOfLifeDate: node10EOL,
               },
               linuxRuntimeSettings: {
@@ -336,6 +428,18 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   linuxFxVersion: 'Node|10',
                 },
                 supportedFunctionsExtensionVersions: ['~2', '~3'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~2',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                  {
+                    version: '~3',
+                    isDeprecated: true,
+                    isDefault: false,
+                  },
+                ],
                 endOfLifeDate: node10EOL,
               },
             },
@@ -368,6 +472,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   use32BitWorkerProcess: true,
                 },
                 supportedFunctionsExtensionVersions: ['~2'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~2',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node8EOL,
               },
             },
@@ -398,6 +509,13 @@ const getNodeStack: (useIsoDateFormat: boolean) => FunctionAppStack = (useIsoDat
                   use32BitWorkerProcess: true,
                 },
                 supportedFunctionsExtensionVersions: ['~1'],
+                supportedFunctionsExtensionVersionsInfo: [
+                  {
+                    version: '~1',
+                    isDeprecated: true,
+                    isDefault: true,
+                  },
+                ],
                 endOfLifeDate: node6EOL,
               },
             },
