--- server/src/stacks/2023-11-01/models/AppStackModel.ts
@@ -1,41 +1,0 @@
-export interface AppStack<T, V> {
-  displayText: string;
-  value: V;
-  majorVersions: AppStackMajorVersion<T>[];
-  preferredOs?: AppStackOs;
-}
-
-export interface AppStackMajorVersion<T> {
-  displayText: string;
-  value: string;
-  minorVersions: AppStackMinorVersion<T>[];
-}
-
-export interface AppStackMinorVersion<T> {
-  displayText: string;
-  value: string;
-  stackSettings: T;
-}
-
-export type AppStackOs = 'linux' | 'windows';
-
-export interface AppInsightsSettings {
-  isSupported: boolean;
-  isDefaultOff?: boolean;
-}
-
-export interface GitHubActionSettings {
-  isSupported: boolean;
-  supportedVersion?: string;
-  notSupportedInCreates?: boolean;
-}
-
-export interface CommonSettings {
-  isPreview?: boolean;
-  isDeprecated?: boolean;
-  isHidden?: boolean;
-  endOfLifeDate?: string;
-  isAutoUpdate?: boolean;
-  isDefault?: boolean;
-  isEarlyAccess?: boolean;
-}
