--- server/src/stacks/2023-11-01/stacks/date-utilities.ts
@@ -1,7 +1,0 @@
-export const getDateString = (date: Date, useIsoDateFormat: boolean) => {
-  if (!date) {
-    return '';
-  }
-
-  return useIsoDateFormat ? date.toISOString() : date.toString();
-};
