--- server/src/stacks/2023-11-01/stacks/web-app-stacks/StaticSite.ts
@@ -1,32 +1,0 @@
-import { WebAppStack } from '../../models/WebAppStackModel';
-
-export const staticSiteStack: WebAppStack = {
-  displayText: 'HTML (Static Content)',
-  value: 'staticsite',
-  preferredOs: 'linux',
-  majorVersions: [
-    {
-      displayText: 'HTML (Static Content)',
-      value: '1',
-      minorVersions: [
-        {
-          displayText: 'HTML (Static Content)',
-          value: '1.0',
-          stackSettings: {
-            linuxRuntimeSettings: {
-              runtimeVersion: 'STATICSITE|1.0',
-              isHidden: true,
-              remoteDebuggingSupported: false,
-              appInsightsSettings: {
-                isSupported: false,
-              },
-              gitHubActionSettings: {
-                isSupported: false,
-              },
-            },
-          },
-        },
-      ],
-    },
-  ],
-};
